Cleaning up synthetic experiments notebook to be more user-friendly. In progress, still need to do proxy rm and semisynthetic data. Need to double check debiased estimate is supposed to be this good and this is not a bug
